+++ NEWS.md
@@ -1,5 +1,8 @@
 # tidyselect (development version)
 
+* `any_of()` generates more informative error if you supply too many
+  arguments (#241).
+
 * `where()` is now exported, like all other select helpers (#201).
 
 * Fixed error when selecting with predicates and `allow_rename =